refactor: extract meta-tool definitions to lean/meta_tools.py (#140)
Move the 345-line _setup_meta_tools body (3 inline tool closures) into
a standalone setup_meta_tools(interface) function in meta_tools.py,
replacing all self. references with interface. The method now delegates
with a 2-line import+call. Also removes unused ValidationError, validate,
and apply_token_limits imports from interface.py.

interface.py: 706 → 366 lines (-340)
meta_tools.py: new file, 372 lines
